Working fetch stage
