+++ src/Backup before assignment 4/membus.hh
@@ -6,19 +6,18 @@
 
 #include <vector>
 #include <deque>
-#include <memory>
 
 class Membus : public SimObject{
 private:
     // ********************************************************************
-    // Packets take time to filter through the membus. This event manages
+    // Packets take time to filter through the membuc. This event manages
     // this process
     // ********************************************************************
     class MembusForwardEvent : public Event{
     private:
         Membus *owner;
     public:
-        MembusForwardEvent(Membus *_owner) : Event(), owner(_owner) {}
+        MembusForwardEvent(Membus *_owner) : event(), owner(_owner) {}
         void process() override { owner->forwardPackets(); }
         const char* description() override { return "Membus Forward Event"; }
 
@@ -28,30 +27,30 @@ private:
     // ********************************************************************
     class MemSidePort : public MasterPort{
     private:
-        Membus *owner;
+    Membus *owner;
         PacketPtr activeRequest;
     public:
-        MemSidePort(Membus * _owner) : owner(_owner) {}
-        bool isBusy() { return (activeRequest); }
-        void recvResp(PacketPtr pkt) override {
+        MemSidePort(Membus * _owner) : Event(), owner(_owner) {}
+        bool isBusy() { return (activeRequest) }
+        void recvReq(Packetptr pkt) override {
             activeRequest = nullptr;
             owner->recvResp(pkt);
         }
-        void sendReq(PacketPtr pkt) override {
+        void sendReq(PacketPtr pkt){
             activeRequest = pkt;
             MasterPort::sendReq(pkt);
         }
     };
     // ********************************************************************
-    // Port that points upwards toward the CPU in the memory hiearchy
+    // Port that points downward toward DRAM in the memory hiearchy
     // ********************************************************************
     class CPUSidePort : public SlavePort{
     private:
-        Membus *owner;
+        membus *owner;
     public:
-        CPUSidePort(Membus *_owner): SlavePort(), owner(_owner) {}
-        virtual void recvReq(PacketPtr pkt) override { owner->recvReq(pkt); }
-        const AddrRange getAddrRange() override { return AddrRange(0,0); }
+        CPUSidePort(membus *_owner): SlavePort(), owner(_owner) {}
+        virtual void recvReq(PacketPtr pkt) override { owner->recvReq(pkt) }
+        const AddrRange getAddrRange() override { return AddrRange(0,0) }
     };
 
     typedef std::pair<Tick, PacketPtr> fwdQType;
@@ -59,24 +58,25 @@ private:
     Tick fwd_time;
     MembusForwardEvent * fwd_event;
     // Attach multiple ports on both the cpu and memory side
-    std::vector<MemSidePort *> memSidePorts; // memory Side connections
-    std::vector<CPUSidePort *> cpuSidePorts; // Cpu side connections
+    std::vector<MemSidePort *> MemSidePorts; // memory Side connections
+    std::vector<CPUSidePort> cpuSidePorts; // Cpu side connections
     std::deque<PacketPtr> packetsWaitingForMemPorts; // Packets that are waiting for memory port to be able to send to memory
     std::deque<fwdQType> packetsWaitingForForward; // Packet forwarding mechanism
 public:
-    Membus(std::shared_ptr<System> sys, Tick forward_time);
-    void initialize() override {}
+    Membus(System *sys, Tick forward_time);
+
     void tryToSend();
     void forwardPackets();
     void recvReq(PacketPtr pkt);
     void recvResp(PacketPtr pkt);
-    MemSidePort *getRequestPort(PacketPtr pkt);
+    MemSideport *getRequestPort(PacketPtr pkt);
     CPUSidePort *getResponsePort(PacketPtr pkt);
 
-    MemSidePort *getMemSidePort(size_t index);
+    MemSideport *getMemSidePort(size_t index);
     CPUSidePort *getCPUSidePort(size_t index);
-    CPUSidePort *getUnboundCPUSidePort();
-    MemSidePort *getUnboundMemSidePort();
+    MemSidePort *getUnboundCPUSidePort();
+    CPUSidePort *getUnboundMemSidePort();
+
 };
