+++ src/Backup before assignment 4/system.hh
@@ -22,7 +22,6 @@ public:
     void printMEQ();
     void removeEvent(); // Removes an event from the MEQ
     void clearMEQ() { MEQ.clear(); }  // Clears the entire MEQ
-    Event *popEvent();
 };
 
 #endif // SYSTEM_H
